+++ packages/pluggableWidgets/datagrid-web/src/Datagrid.editorPreview.tsx
@@ -5,6 +5,7 @@ import { Table, TableColumn } from "./components/Table";
 import { parseStyle } from "@mendix/piw-utils-internal";
 import { Selectable } from "mendix/preview/Selectable";
 import { ObjectItem, GUID } from "mendix";
+import classNames from "classnames";
 
 interface PreviewProps extends Omit<DatagridPreviewProps, "class"> {
     className: string;
@@ -32,7 +33,8 @@ export function preview(props: PreviewProps): ReactElement {
                       hidable: "no",
                       size: 1,
                       sortable: false,
-                      alignment: "left"
+                      alignment: "left",
+                      wrapText: false
                   }
               ];
 
@@ -57,7 +59,7 @@ export function preview(props: PreviewProps): ReactElement {
             cellRenderer={useCallback(
                 (renderWrapper, _, columnIndex) => {
                     const column = columns[columnIndex];
-                    const className = column.alignment ? `align-column-${column.alignment}` : "";
+                    const className = classNames(`align-column-${column.alignment}`, { "wrap-text": column.wrapText });
                     let content;
                     switch (column.showContentAs) {
                         case "attribute":