+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+- We changed the icons from front-awesome to be pure SVG.
+
 ## [2.2.0] - 2021-12-23
 
 ### Added