+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.2.1] - 2022-1-6
+
 ### Changed
 - We changed the icons from front-awesome to be pure SVG.